feat : reactStrict 모드 해제
