+++ data/TestData.ts
@@ -1,0 +1,14 @@
+export const myEventsList = [
+	{
+		title: '수원 여행 가자',
+		start: new Date(2023, 9, 16, 12, 6, 0),
+		end: new Date(2023, 9, 18, 12, 6, 0),
+	},
+	{
+		title: '평택 여행 가자',
+		start: new Date(2023, 9, 18, 12, 6, 0),
+		end: new Date(2023, 9, 20, 12, 6, 0),
+	},
+];
+
+export const myItemTitle = ['테스트1', '테스트2', '테스트3'];
