+++ package.json
@@ -17,6 +17,7 @@
 		"@react-spring/web": "^9.7.3",
 		"@tanstack/react-query": "^4.35.3",
 		"@tanstack/react-query-devtools": "^4.35.3",
+		"@types/react-beautiful-dnd": "^13.1.6",
 		"@vercel/blob": "^0.13.0",
 		"@vercel/kv": "^0.2.3",
 		"autoprefixer": "latest",
@@ -37,6 +38,7 @@
 		"postcss": "latest",
 		"prettier": "^3.0.3",
 		"react": "latest",
+		"react-beautiful-dnd": "^13.1.1",
 		"react-big-calendar": "^1.8.4",
 		"react-calendar": "^4.6.0",
 		"react-dom": "latest",
